% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_mate_finder.R
\name{fs_mate_finder}
\alias{fs_mate_finder}
\title{Find fullsib mates}
\usage{
fs_mate_finder(ped)
}
\arguments{
\item{ped}{: A pedigree \code{data.frame}. The first three columns (ID, SIRE, DAM) are used.}
}
\value{
fs_mates : A \code{data.frame} with two columns (SIRE, DAM) representing fullsib mates.
}
\description{
Find fullsib matings in the pedigree
}
\examples{
ped = data.frame(ID=1:7, SIRE=c(0,0,1,0,3,3,5), DAM=c(0,0,0,2,4,4,6))
fs_mate_finder(ped)

}
