% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_mate_finder.R
\name{hs_mate_finder}
\alias{hs_mate_finder}
\title{Find halfsib mates}
\usage{
hs_mate_finder(ped)
}
\arguments{
\item{ped}{: A pedigree \code{data.frame}. The first three columns (ID, SIRE, DAM) are used.}
}
\value{
hs_mates : A \code{data.frame} with two columns (SIRE, DAM) representing halfsib mates.
}
\description{
Find halfsib matings in the pedigree
}
\examples{
ped = data.frame(ID=1:7, SIRE=c(0,0,1,1,0,3,5), DAM=c(0,0,2,2,2,4,4))
hs_mate_finder(ped)

}
