% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedSimulate.R
\docType{package}
\name{pedSimulte-package}
\alias{pedSimulte-package}
\title{Pedigree, genetic merit and phenotype simulation}
\description{
An R package for simulating a pedigree with genetic merits and phenotypes, starting from a base population (generation 0) or an existing pedigree.
The pedigree depth and design can be chosen by the values provided to the arguments of the simulation function.
}
\details{
Starting from a base population with a user-defined size and equal number of males and females,
next generations are simulated for the user-defined littersize and number of generations.
No selection (natural or artificial) and non-random mating is applied to this population.
Alternatively, the simulation can be started from an existing pedigree.
Natural (mortality) and artificial selection are applied to the next generations.
Different generation overlap, and selection intensities and patterns can be applied to males and females.
Selected individuals are mated randomly or assortatively/disassortatively, 
based on different selection criteria applied to males and females, independently.
Performance and genetic merit of individuals are simulated using the basic rules of quantitative genetics.
The performance (P) of an individual is influenced by genetic (A) and environmental (E) effects.
Thus, P = A + E, and Var(P) = Var(A) + Var(E).
The additive genetic merit (A) of an individual is the average of its parents' additive genetic merits
(PA = (A\ifelse{html}{\out{<sub>sire</sub>}}{\eqn{_{sire}}} + A\ifelse{html}{\out{<sub>dam</sub>}}{\eqn{_{dam}}})/2)
plus the Mendelian Sampling term due to the sampling of alleles passed from the parent to the offspring.
The Mendelian Sampling variance is half of Var(A) in the base population.
Because there is no provided information for environmental effects, the environment effect is
assigned to individuals from a normal distribution of random numbers (E ~ N(0, \strong{I}Var(E))).
The package also provides functions to identify halfsib, fullsib and parent-progeny matings in the pedigree.
}
\author{
Mohammad Ali Nilforooshan \email{m.a.nilforooshan@gmail.com}
}
