% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_mate_finder.R
\name{pp_mate_finder}
\alias{pp_mate_finder}
\title{Find parent-progeny mates}
\usage{
pp_mate_finder(ped)
}
\arguments{
\item{ped}{: A pedigree \code{data.frame}. The first three columns (ID, SIRE, DAM) are used.}
}
\value{
pp_mates : A \code{data.frame} with two columns (SIRE, DAM) representing parent-progeny mates.
}
\description{
Find parent-progeny matings in the pedigree
}
\examples{
ped = data.frame(ID=1:4, SIRE=c(0,0,1,1), DAM=c(0,0,2,3))
pp_mate_finder(ped)

}
